Added a common superclass for ACL commands handling command name validation

+++ src/main/java/dev/vrba/pyro/discord/modules/core/commands/acl/ResetCommand.java
@@ -5,18 +5,17 @@ import dev.vrba.pyro.discord.commands.acl.ACLEntriesRepository;
 import dev.vrba.pyro.discord.commands.acl.ACLEntry;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 @Component
-public class ResetCommand implements Command {
-
-    private final ACLEntriesRepository repository;
+public class ResetCommand extends ACLCommand implements Command {
 
     @Autowired
-    public ResetCommand(@NotNull final ACLEntriesRepository repository) {
-        this.repository = repository;
+    public ResetCommand(@NotNull final ACLEntriesRepository repository, @NotNull final ApplicationContext context) {
+        super(repository, context);
     }
 
     @Override
@@ -53,6 +52,15 @@ public class ResetCommand implements Command {
 
         long guildId = context.getEvent().getGuild().getIdLong();
         final String name = context.getArguments().get(0);
+
+        if (!isTargetCommandValid(name)) {
+            CommandUtils.sendError(
+                    context,
+                    "Unknown or invalid command `" + name + "`",
+                    "Make sure the target command exists and that its execution is ACL-dependent."
+            );
+            return;
+        }
 
         final List<ACLEntry> deleted = repository.deleteAllByGuildIdAndCommand(guildId, name);
